Various code style fixes 11/01/2021 | 21w44a1 | 0.7.93

+++ Games/Tests/src/Vulkan/VulkanIcoSphereTests.cpp
@@ -98,7 +98,7 @@ void VulkanIcoSphereTests::OnUpdate(const TRAP::Utils::TimeStep&)
 	m_vertexBuffer->Use();
 	m_indexBuffer->Use();
 	m_shader->Use();
-	TRAP::Graphics::RenderCommand::DrawIndexed(m_icosphereIndices.size());
+	TRAP::Graphics::RenderCommand::DrawIndexed(static_cast<uint32_t>(m_icosphereIndices.size()));
 
 	//Simple performance metrics
 	if (m_fpsTimer.Elapsed() >= 5.0f) //Output Every 5 Seconds
