+++ TRAP/src/Graphics/Renderer.cpp
@@ -26,7 +26,7 @@ void TRAP::Graphics::Renderer::Init()
 	TP_PROFILE_FUNCTION();
 
 	s_maxDrawCalls = RendererAPI::GPUSettings.MaxUniformBufferRange /
-	                 UniformBuffer::CalculateAlignedSize(sizeof(Math::Mat4));
+	                 static_cast<uint32_t>(UniformBuffer::CalculateAlignedSize(sizeof(Math::Mat4)));
 
 	s_uniformBuffer = TRAP::Graphics::UniformBuffer::Create(s_sceneData.get(), sizeof(SceneData),
 	 														TRAP::Graphics::UpdateFrequency::Dynamic);
